requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-huggingface>=0.0.1
langchain-text-splitters>=0.0.1
huggingface-hub>=0.20.0
python-dotenv>=1.0.0
requests>=2.31.0
python-docx>=1.1.0
PyPDF2>=3.0.0
reportlab>=4.0.0
openpyxl>=3.1.0
scikit-learn>=1.3.0
numpy>=1.24.0
